--- bcs-app/backend/apps/datalog/constant.py
@@ -1,21 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# Tencent is pleased to support the open source community by making 蓝鲸智云PaaS平台社区版 (BlueKing PaaS Community Edition) available.
-# Copyright (C) 2017-2019 THL A29 Limited, a Tencent company. All rights reserved.
-# Licensed under the MIT License (the "License"); you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://opensource.org/licenses/MIT
-#
-# Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
-# an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
-# specific language governing permissions and limitations under the License.
-#
-
-
-# 数据平台申请dataid的名称前缀
-STANDARD_LOG_PREFIX_KEY = 'STANDARD_LOG_PREFIX_KEY'
-NON_STANDARD_LOG_PREFIX_KEY = 'NON_STANDARD_LOG_PREFIX_KEY'
-
-STANDARD_LOG_PREFIX_VALUE = 'bcs_std_log_'
-NON_STANDARD_LOG_PREFIX_VALUE = 'bcs_non_std_log_'
