+++ bcs-app/backend/apps/network/views/service.py
@@ -23,7 +23,6 @@ from rest_framework.response import Response
 
 from backend.accounts import bcs_perm
 from backend.activity_log import client as activity_client
-from backend.apps import constants
 from backend.apps import utils as app_utils
 from backend.apps.application.base_views import BaseAPI
 from backend.apps.application.constants import DELETE_INSTANCE, SOURCE_TYPE_MAP
@@ -58,8 +57,9 @@ from backend.apps.instance.generator import (
 )
 from backend.apps.instance.models import InstanceConfig
 from backend.apps.instance.utils_pub import get_cluster_version
+from backend.apps.network.ext_routes import delete_svc_extended_routes, get_svc_extended_routes
 from backend.apps.network.serializers import BatchResourceSLZ
-from backend.apps.network.utils_bk import delete_svc_extended_routes, get_svc_access_info, get_svc_extended_routes
+from backend.apps.network.utils import get_svc_access_info
 from backend.components.bcs import k8s, mesos
 from backend.resources.namespace.constants import K8S_PLAT_NAMESPACE, K8S_SYS_NAMESPACE
 from backend.utils.errcodes import ErrorCode
